x86-64: Display directly-accessed static-symbols in disassembler.
  * When accessing the value slot of a static-symbol, the
compiler emits an EA with the absolute address of the slot.
When this EA is disassembled, it shows as a meaningless
address.

  * When disassembling such an EA, check for such cases by
subtracting NIL and checking for a NIL-indexed symbol slot
reference.

  * In SB-DISASSEM:GROK-SYMBOL-SLOT-REF, be more clever
when calling SB-KERNEL:MAKE-LISP-OBJ, as this function is
now called speculatively as well as when the address is
known to be a symbol slot address.