x86: Display referenced immediate constants in disassembler.
  * When loading a certain class of immediate constants into
registers, such as static-symbols, the compiler emits a MOV
instruction with the constant as an immediate.  When this MOV
is disassembled, it showed as a meaningless number.

  * Define a print-* function that tries to turn the immediate
value into a boxed lisp object, and displays it as a note if
successful.

  * Spam the new print function across all of the
immediate-data arg-formats, as the one used for x86-64 isn't
the right one for x86 (and, bonus, we catch unbound-markers
this way, and leave the door open for other tags).