x86: Display directly-accessed static-symbols in disassembler.
  * When accessing the value slot of a static-symbol, the
compiler emits an EA with the absolute address of the slot.
When this EA is disassembled, it shows as a meaningless
address.

  * When disassembling such an EA, check for such cases by
bounds-checking against static space, then subtracting NIL
and checking for a NIL-indexed symbol slot reference.

  * The bounds-check should probably be in the x86-64
version as well, but it didn't fail under light testing,
and that version was written first.